# JongSnamFootball.Entities/Dtos/GrahpDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace JongSnamFootball.Entities.Dtos
{
    public class GrahpDto
    {
        public int? Days { get; set; }
        public int? Months { get; set; }
        public int? Years { get; set; }

    }
}