# Tools/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "JongSnam.Services",
    "version": "v1"
  },
  "paths": {
...
    "/api/Reservation/{userId}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "operationId": "GetYourReservation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReservationDtoBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation/{userId}/{month}/month": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "operationId": "GetGraphReservation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "month",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrahpDtoBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation/{userId}/{year}/year": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "operationId": "GraphYearReservation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "year",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GrahpDtoBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation/Search": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "operationId": "GetReservationBySearch",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StartTime.Year",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StartTime.Month",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StartTime.Day",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StartTime.Hour",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StartTime.Minute",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StartTime.Second",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StopTime.Year",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StopTime.Month",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StopTime.Day",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StopTime.Hour",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StopTime.Minute",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "StopTime.Second",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "UserName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "StoreName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReservationDtoBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation/{Id}/Detail": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "operationId": "GetShowDetailYourReservation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReservationDetailDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation/ApprovalStatus": {
      "put": {
        "tags": [
          "Reservation"
        ],
        "operationId": "UpdateApprovalStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReservationApprovalRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "operationId": "CreateReservation",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReservationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Reservation/{id}/DeleteReservation": {
      "put": {
        "tags": [
          "Reservation"
        ],
        "operationId": "DeleteReservation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Review/{storeId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "operationId": "GetReviewByStoreId",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SumaryRatingDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Review": {
      "post": {
        "tags": [
          "Review"
        ],
        "operationId": "AddReview",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProblemsDetailDto": {
      "type": "object",
      "properties": {
        "detail": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        },
        "instance": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "errorCode": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "ReservationDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "approvalStatus": {
          "type": "boolean"
        },
        "userName": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "contactMobile": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "stopTime": {
          "format": "date-time",
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ReservationDtoBasePagingDto": {
      "type": "object",
      "properties": {
        "currentPage": {
          "format": "int32",
          "type": "integer"
        },
        "totalPage": {
          "format": "int32",
          "type": "integer"
        },
        "collection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReservationDto"
          }
        }
      }
    },
    "GrahpDto": {
      "type": "object",
      "properties": {
        "days": {
          "format": "int32",
          "type": "integer"
        },
        "months": {
          "format": "int32",
          "type": "integer"
        },
        "years": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "GrahpDtoBasePagingDto": {
      "type": "object",
      "properties": {
        "currentPage": {
          "format": "int32",
          "type": "integer"
        },
        "totalPage": {
          "format": "int32",
          "type": "integer"
        },
        "collection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GrahpDto"
          }
        }
      }
    },
    "PaymentModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "updatedDate": {
          "format": "date-time",
          "type": "string"
        },
        "reservationId": {
          "format": "int32",
          "type": "integer"
        },
        "image": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "isFullAmount": {
          "type": "boolean"
        }
      }
    },
    "ReservationDetailDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "approvalStatus": {
          "type": "boolean"
        },
        "imageProfile": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "contactMobile": {
          "type": "string"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "stopTime": {
          "format": "date-time",
          "type": "string"
        },
        "fieldName": {
          "type": "string"
        },
        "pricePerHour": {
          "format": "double",
          "type": "number"
        },
        "isFullAmount": {
          "type": "boolean"
        },
        "paymentModel": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PaymentModel"
          }
        },
        "amountForPay": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "ReservationApprovalRequest": {
      "type": "object",
      "properties": {
        "approvalStatus": {
          "type": "boolean"
        }
      }
    },
    "ReservationRequest": {
      "type": "object",
      "properties": {
        "storeId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "fieldId": {
          "format": "int32",
          "type": "integer"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "stopTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ReviewDto": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "SumaryRatingDto": {
      "type": "object",
      "properties": {
        "currentPage": {
          "format": "int32",
          "type": "integer"
        },
        "totalPage": {
          "format": "int32",
          "type": "integer"
        },
        "collection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewDto"
          }
        },
        "totalFive": {
          "format": "int32",
          "type": "integer"
        },
        "totalFour": {
          "format": "int32",
          "type": "integer"
        },
        "totalThree": {
          "format": "int32",
          "type": "integer"
        },
        "totalTwo": {
          "format": "int32",
          "type": "integer"
        },
        "totalOne": {
          "format": "int32",
          "type": "integer"
        },
        "summaryRating": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "ReviewRequest": {
      "required": [
        "message",
        "rating",
        "storeId",
        "userId"
      ],
      "type": "object",
      "properties": {
        "storeId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer 12345abcdef"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}